bitonic sorts for one and two vector work, for two float vectors not yet implemented

+++ geometrical_objects/src/print_functions.hpp
@@ -143,7 +143,7 @@ inline void print_avx(__m256d reg, std::string name = "") {
 
 inline void print_avx(__m256 reg, std::string name = "") {
     std::cout << name << " ";
-    double *r = (double *)&reg;
+    float *r = (float *)&reg;
     std::cout << *r << " " << *(r + 1) << " " << *(r + 2) << " " << *(r + 3)
               << " " << *(r + 4) << " " << *(r + 5) << " " << *(r + 6) << " "
               << *(r + 7) << std::endl;
